ATR-304: added base class and architecture for suppression command. Added template method for suppression.
