+++ src/Acuminator/Acuminator.Vsix/Commands/DiagnosticSuppression/Diagnostic Service/DiagnosticData.cs
