+++ src/Acuminator/Acuminator.Vsix/Commands/DiagnosticSuppression/Suppression File/ProjectSuppressionFileGenerator.cs
